+++ frontend/src/components/EnvironmentForm/index.ts
@@ -1,0 +1,7 @@
+import Buttons from "./Buttons.vue";
+import EnvironmentForm from "./EnvironmentForm.vue";
+import Form from "./Form.vue";
+
+export * from "./context";
+export { EnvironmentForm, Form, Buttons };
+export default EnvironmentForm;
